feat: batch import/experimental hpc + fixes

+++ wipp-backend-data/src/main/java/gov/nist/itl/ssd/wipp/backend/data/imagescollection/images/ImageConversionService.java
@@ -15,6 +15,7 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Optional;
 import java.nio.file.Paths;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -67,13 +68,24 @@ public class ImageConversionService extends FileUploadBase{
 
 	@PostConstruct
 	public void instantiateOmeConverter() {
-		omeConverterExecutor = Executors.newFixedThreadPool(
-				appConfig.getOmeConverterThreads());
+		omeConverterExecutor = Executors.newFixedThreadPool(1);
+
+		// Load security context for system operations
+		SecurityUtils.runAsSystem();
 
 		// Resume any interrupted conversion
-		SecurityUtils.runAsSystem();
 		imageRepository.findByImporting(true)
-		.forEach(this::submitImageToExtractor);
+                .forEach(image -> {
+                    Optional<ImagesCollection> coll = imagesCollectionRepository.findById(image.getImagesCollection());
+					if (coll.isPresent() && ImagesCollection.ImagesCollectionImportMethod.BACKEND_IMPORT.equals(
+							coll.get().getImportMethod())) {
+						File sourceDir = new File(appConfig.getLocalImportFolder(), coll.get().getSourceBackendImport());
+						this.submitImageToExtractor(image, sourceDir, false);
+					} else {
+						this.submitImageToExtractor(image);
+					}
+                });
+		// Clear security context after system operations
 		SecurityContextHolder.clearContext();
 	}
 
@@ -84,10 +96,10 @@ public class ImageConversionService extends FileUploadBase{
 	
 	public void submitImageToExtractor(Image image) {
 		File tempUploadDir = getTempUploadDir(image.getImagesCollection());
-		this.submitImageToExtractor(image, tempUploadDir);
+		this.submitImageToExtractor(image, tempUploadDir, true);
 	}
 
-	public void submitImageToExtractor(Image image, File sourceDir) {
+	public void submitImageToExtractor(Image image, File sourceDir, boolean deleteSourceImage) {
 		String collectionId = image.getImagesCollection();
 		ImagesCollection imgCollection = imagesCollectionRepository.findById(collectionId).orElse(null);
 		// if images collection not found, image should be deleted to avoid inconsistent state
@@ -116,17 +128,19 @@ public class ImageConversionService extends FileUploadBase{
 		Path outputPath = new File(uploadDir, outputFileName).toPath();
 
 		omeConverterExecutor.submit(() -> doSubmit(
-				collectionId, image, outputFileName, tempPath, outputPath));
+				collectionId, image, outputFileName, tempPath, outputPath, deleteSourceImage));
 	}
 
 	public void doSubmit(String collectionId, Image image, String outputFileName,
-			Path tempPath, Path outputPath) {
+			Path tempPath, Path outputPath, boolean deleteSourceImage) {
 		try {
 			LOG.log(Level.INFO,
 					"Starting extracting image {0} of collection {1}",
 					new Object[]{image.getFileName(), collectionId});
 			convertToTiledOmeTiff(tempPath, outputPath);
-			Files.delete(tempPath);
+			if (deleteSourceImage) {
+				Files.delete(tempPath);
+			}
 			image.setFileName(outputFileName);
 			image.setFileSize(getPathSize(outputPath));
 			image.setImporting(false);
@@ -147,7 +161,7 @@ public class ImageConversionService extends FileUploadBase{
 		}
 	}
 	
-	public static void convertToTiledOmeTiff(Path inputFile, Path outputFile) throws DependencyException, FormatException,
+	public void convertToTiledOmeTiff(Path inputFile, Path outputFile) throws DependencyException, FormatException,
 			IOException, ServiceException {
 		String omeTiffOutputName = outputFile.toString();
 		String omeZarrOutputName = omeTiffOutputName.substring(0, omeTiffOutputName.lastIndexOf('.')) + ".zarr";
@@ -159,6 +173,8 @@ public class ImageConversionService extends FileUploadBase{
 	    			omeZarrOutputName,
 	    			"--tile-height", String.valueOf(CoreConfig.TILE_SIZE),
 	    			"--tile-width", String.valueOf(CoreConfig.TILE_SIZE),
+					"--max-workers", String.valueOf(appConfig.getOmeConverterThreads()),
+					"--log-level", "ERROR"
 	    	};
 	    	CommandLine.call(new Converter(), converterArgs);
 
@@ -168,6 +184,8 @@ public class ImageConversionService extends FileUploadBase{
 	    			outputFile.toString(),
 	    			"--rgb",
 	    			"--compression", "LZW",
+					"--max_workers", String.valueOf(appConfig.getOmeConverterThreads()),
+					"--log-level", "ERROR"
 	    	};
 	    	CommandLine.call(new PyramidFromDirectoryWriter(), converterPyrArgs);
